examples/python-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.34.1
pydantic>=2.11.2
python-multipart==0.0.20
ag-ui-protocol
python-dotenv>=1.0.0
-e ../../packages/agentkit